refactor: add custom exceptions
