+++ app/src/main/resources/application.properties
@@ -1,9 +1,11 @@
 spring.application.name=MENA
 
-spring.datasource.url=${DATABASE_URL}
-spring.datasource.username=${DATABASE_USER_NAME}
-spring.datasource.password=${DATABASE_PASSWORD}
+spring.datasource.url=jdbc:postgresql://localhost:5432/mena
+spring.datasource.username=postgres
+spring.datasource.password=postgres
 
-jwt.secret-key=${JWT_SECRET}
+server.port=8085
 
-spring.jpa.hibernate.ddl-auto=validate
+jwt.secret-key=fb798677db0384cc411554a608406e0e35779a43873568d90093981e8c9a6afc
+
+spring.jpa.hibernate.ddl-auto=update
