+++ trends/src/main/kotlin/net/thechance/trends/api/controller/exception/ReelNotFoundException.kt
@@ -1,0 +1,3 @@
+package net.thechance.trends.api.controller.exception
+
+class ReelNotFoundException: Exception("Reel not found")
