+++ trends/src/main/kotlin/net/thechance/trends/service/ReelsService.kt
@@ -1,5 +1,6 @@
 package net.thechance.trends.service
 
+import net.thechance.trends.api.controller.exception.ReelNotFoundException
 import net.thechance.trends.repository.ReelsRepository
 import org.springframework.stereotype.Service
 import java.util.*
@@ -12,7 +13,7 @@ class ReelsService(
     fun deleteReelById(id: UUID, currentUserId: UUID) {
         val isReelExists = reelsRepository.existsByIdAndOwnerId(id , currentUserId)
 
-        if (!isReelExists) throw NoSuchElementException("This reel cannot be found")
+        if (!isReelExists) throw ReelNotFoundException()
 
         reelsRepository.deleteById(id)
     }
